Moving custom functions from app-frontend.ts into custom.ts
